gfk grappelli widget: adjusted duplication fix
There are some edge cases with the fix in 125890c for generic foreign
key widgets. The issue seems to be that grappelli switched to an init /
update style implementation for that widget in
static/grappelli/js/jquery.grp_related_generic.js -- the only way to
resolve it is to copy lines 56,57 of the update_lookup function from
there which just removes these elements before re-adding them.

Closes Pull Request #41